refactor(weights): remove cog inspect command
The hidden cog inspect command was added during early managed-weights
development to debug bundle pushes. It was never designed for general
release: no docs, no tests beyond a single integration script, and the
output format (text and JSON variants, with a separate raw streaming
mode) accreted features without a coherent target audience.

Removing it lets us also drop pkg/model/index.go (the Index/IndexManifest
parse types only inspect consumed) and a small block in resolver.go that
populated those types when loading bundles. The push-side IndexBuilder
in index_factory.go is unrelated and stays.

PlatformUnknown moves next to Platform in artifact_image.go, where it
belongs structurally. TestModel_IsBundle moves to model_test.go since
index_test.go went away.
