+++ pkg/cli/root.go
@@ -47,7 +47,6 @@ https://github.com/replicate/cog`,
 		newDebugCommand(),
 		newDoctorCommand(),
 		newInitCommand(),
-		newInspectCommand(),
 		newLoginCommand(),
 		newPredictCommand(),
 		newPushCommand(),
